chore: add runtime image development baseline

+++ services/document-updater/test/acceptance/js/DeletingAProjectTests.js
@@ -1,5 +1,13 @@
+// TODO: This file was created by bulk-decaffeinate.
+// Fix any style issues and re-enable lint.
+/*
+ * decaffeinate suggestions:
+ * DS101: Remove unnecessary use of Array.from
+ * DS207: Consider shorter variations of null checks
+ * Full docs: https://github.com/decaffeinate/decaffeinate/blob/master/docs/suggestions.md
+ */
 const sinon = require('sinon')
-const { setTimeout } = require('node:timers/promises')
+const async = require('async')
 
 const MockProjectHistoryApi = require('./helpers/MockProjectHistoryApi')
 const MockWebApi = require('./helpers/MockWebApi')
@@ -7,7 +15,7 @@ const DocUpdaterClient = require('./helpers/DocUpdaterClient')
 const DocUpdaterApp = require('./helpers/DocUpdaterApp')
 
 describe('Deleting a project', function () {
-  beforeEach(async function () {
+  beforeEach(function (done) {
     let docId0, docId1
     this.project_id = DocUpdaterClient.randomId()
     this.docs = [
@@ -42,27 +50,45 @@ describe('Deleting a project', function () {
         updatedLines: ['four', 'four and a half', 'five', 'six'],
       },
     ]
-    for (const doc of this.docs) {
+    for (const doc of Array.from(this.docs)) {
       MockWebApi.insertDoc(this.project_id, doc.id, {
         lines: doc.lines,
         version: doc.update.v,
       })
     }
 
-    await DocUpdaterApp.ensureRunning()
+    DocUpdaterApp.ensureRunning(done)
   })
 
   describe('without updates', function () {
-    beforeEach(async function () {
+    beforeEach(function (done) {
       sinon.spy(MockWebApi, 'setDocument')
       sinon.spy(MockProjectHistoryApi, 'flushProject')
 
-      for (const doc of this.docs) {
-        await DocUpdaterClient.preloadDoc(this.project_id, doc.id)
-      }
-      await setTimeout(200)
-      const res = await DocUpdaterClient.deleteProject(this.project_id)
-      this.statusCode = res.status
+      async.series(
+        this.docs.map(doc => {
+          return callback => {
+            DocUpdaterClient.preloadDoc(this.project_id, doc.id, error => {
+              callback(error)
+            })
+          }
+        }),
+        error => {
+          if (error != null) {
+            throw error
+          }
+          setTimeout(() => {
+            DocUpdaterClient.deleteProject(
+              this.project_id,
+              (error, res, body) => {
+                if (error) return done(error)
+                this.statusCode = res.statusCode
+                done()
+              }
+            )
+          }, 200)
+        }
+      )
     })
 
     afterEach(function () {
@@ -78,18 +104,32 @@ describe('Deleting a project', function () {
       MockWebApi.setDocument.should.not.have.been.called
     })
 
-    it('should need to reload the docs if read again', async function () {
+    it('should need to reload the docs if read again', function (done) {
       sinon.spy(MockWebApi, 'getDocument')
-      for (const doc of this.docs) {
-        MockWebApi.getDocument
-          .calledWith(this.project_id, doc.id)
-          .should.equal(false)
-        await DocUpdaterClient.getDoc(this.project_id, doc.id)
-        MockWebApi.getDocument
-          .calledWith(this.project_id, doc.id)
-          .should.equal(true)
-      }
-      MockWebApi.getDocument.restore()
+      async.series(
+        this.docs.map(doc => {
+          return callback => {
+            MockWebApi.getDocument
+              .calledWith(this.project_id, doc.id)
+              .should.equal(false)
+            DocUpdaterClient.getDoc(
+              this.project_id,
+              doc.id,
+              (error, res, returnedDoc) => {
+                if (error) return done(error)
+                MockWebApi.getDocument
+                  .calledWith(this.project_id, doc.id)
+                  .should.equal(true)
+                callback()
+              }
+            )
+          }
+        }),
+        () => {
+          MockWebApi.getDocument.restore()
+          done()
+        }
+      )
     })
 
     it('should flush each doc in project history', function () {
@@ -100,16 +140,44 @@ describe('Deleting a project', function () {
   })
 
   describe('with documents which have been updated', function () {
-    beforeEach(async function () {
+    beforeEach(function (done) {
       sinon.spy(MockWebApi, 'setDocument')
       sinon.spy(MockProjectHistoryApi, 'flushProject')
-      for (const doc of this.docs) {
-        await DocUpdaterClient.preloadDoc(this.project_id, doc.id)
-        await DocUpdaterClient.sendUpdate(this.project_id, doc.id, doc.update)
-      }
-      await setTimeout(200)
-      const res = await DocUpdaterClient.deleteProject(this.project_id)
-      this.statusCode = res.status
+
+      async.series(
+        this.docs.map(doc => {
+          return callback => {
+            DocUpdaterClient.preloadDoc(this.project_id, doc.id, error => {
+              if (error != null) {
+                return callback(error)
+              }
+              DocUpdaterClient.sendUpdate(
+                this.project_id,
+                doc.id,
+                doc.update,
+                error => {
+                  callback(error)
+                }
+              )
+            })
+          }
+        }),
+        error => {
+          if (error != null) {
+            throw error
+          }
+          setTimeout(() => {
+            DocUpdaterClient.deleteProject(
+              this.project_id,
+              (error, res, body) => {
+                if (error) return done(error)
+                this.statusCode = res.statusCode
+                done()
+              }
+            )
+          }, 200)
+        }
+      )
     })
 
     afterEach(function () {
@@ -122,25 +190,39 @@ describe('Deleting a project', function () {
     })
 
     it('should send each document to the web api', function () {
-      for (const doc of this.docs) {
+      Array.from(this.docs).map(doc =>
         MockWebApi.setDocument
           .calledWith(this.project_id, doc.id, doc.updatedLines)
           .should.equal(true)
-      }
+      )
     })
 
-    it('should need to reload the docs if read again', async function () {
+    it('should need to reload the docs if read again', function (done) {
       sinon.spy(MockWebApi, 'getDocument')
-      for (const doc of this.docs) {
-        MockWebApi.getDocument
-          .calledWith(this.project_id, doc.id)
-          .should.equal(false)
-        await DocUpdaterClient.getDoc(this.project_id, doc.id)
-        MockWebApi.getDocument
-          .calledWith(this.project_id, doc.id)
-          .should.equal(true)
-      }
-      MockWebApi.getDocument.restore()
+      async.series(
+        this.docs.map(doc => {
+          return callback => {
+            MockWebApi.getDocument
+              .calledWith(this.project_id, doc.id)
+              .should.equal(false)
+            DocUpdaterClient.getDoc(
+              this.project_id,
+              doc.id,
+              (error, res, returnedDoc) => {
+                if (error) return done(error)
+                MockWebApi.getDocument
+                  .calledWith(this.project_id, doc.id)
+                  .should.equal(true)
+                callback()
+              }
+            )
+          }
+        }),
+        () => {
+          MockWebApi.getDocument.restore()
+          done()
+        }
+      )
     })
 
     it('should flush each doc in project history', function () {
@@ -151,18 +233,44 @@ describe('Deleting a project', function () {
   })
 
   describe('with the background=true parameter from realtime and no request to flush the queue', function () {
-    beforeEach(async function () {
+    beforeEach(function (done) {
       sinon.spy(MockWebApi, 'setDocument')
       sinon.spy(MockProjectHistoryApi, 'flushProject')
-      for (const doc of this.docs) {
-        await DocUpdaterClient.preloadDoc(this.project_id, doc.id)
-        await DocUpdaterClient.sendUpdate(this.project_id, doc.id, doc.update)
-      }
-      await setTimeout(200)
-      const res = await DocUpdaterClient.deleteProjectOnShutdown(
-        this.project_id
+
+      async.series(
+        this.docs.map(doc => {
+          return callback => {
+            DocUpdaterClient.preloadDoc(this.project_id, doc.id, error => {
+              if (error != null) {
+                return callback(error)
+              }
+              DocUpdaterClient.sendUpdate(
+                this.project_id,
+                doc.id,
+                doc.update,
+                error => {
+                  callback(error)
+                }
+              )
+            })
+          }
+        }),
+        error => {
+          if (error != null) {
+            throw error
+          }
+          setTimeout(() => {
+            DocUpdaterClient.deleteProjectOnShutdown(
+              this.project_id,
+              (error, res, body) => {
+                if (error) return done(error)
+                this.statusCode = res.statusCode
+                done()
+              }
+            )
+          }, 200)
+        }
       )
-      this.statusCode = res.status
     })
 
     afterEach(function () {
@@ -184,21 +292,45 @@ describe('Deleting a project', function () {
   })
 
   describe('with the background=true parameter from realtime and a request to flush the queue', function () {
-    beforeEach(async function () {
+    beforeEach(function (done) {
       sinon.spy(MockWebApi, 'setDocument')
       sinon.spy(MockProjectHistoryApi, 'flushProject')
-      for (const doc of this.docs) {
-        await DocUpdaterClient.preloadDoc(this.project_id, doc.id)
-        await DocUpdaterClient.sendUpdate(this.project_id, doc.id, doc.update)
-      }
-      await setTimeout(200)
-      const res = await DocUpdaterClient.deleteProjectOnShutdown(
-        this.project_id
+
+      async.series(
+        this.docs.map(doc => {
+          return callback => {
+            DocUpdaterClient.preloadDoc(this.project_id, doc.id, error => {
+              if (error != null) {
+                return callback(error)
+              }
+              DocUpdaterClient.sendUpdate(
+                this.project_id,
+                doc.id,
+                doc.update,
+                error => {
+                  callback(error)
+                }
+              )
+            })
+          }
+        }),
+        error => {
+          if (error != null) {
+            throw error
+          }
+          setTimeout(() => {
+            DocUpdaterClient.deleteProjectOnShutdown(
+              this.project_id,
+              (error, res, body) => {
+                if (error) return done(error)
+                this.statusCode = res.statusCode
+                // after deleting the project and putting it in the queue, flush the queue
+                setTimeout(() => DocUpdaterClient.flushOldProjects(done), 2000)
+              }
+            )
+          }, 200)
+        }
       )
-      this.statusCode = res.status
-      // after deleting the project and putting it in the queue, flush the queue
-      await setTimeout(2000)
-      await DocUpdaterClient.flushOldProjects()
     })
 
     afterEach(function () {
@@ -211,11 +343,11 @@ describe('Deleting a project', function () {
     })
 
     it('should send each document to the web api', function () {
-      for (const doc of this.docs) {
+      Array.from(this.docs).map(doc =>
         MockWebApi.setDocument
           .calledWith(this.project_id, doc.id, doc.updatedLines)
           .should.equal(true)
-      }
+      )
     })
 
     it('should flush to project history', function () {
